Move FavoriteListRepository to its own file
