+++ src/app/elitefour/modals/export-data-modal/export-data-modal.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
-import {FavoriteList, FavoriteListsRepository} from '../../backend/favorite-list-interfaces';
+import {FavoriteList} from '../../backend/favorite-list-interfaces';
+import {FavoriteListsRepository} from '../../backend/favorite-lists-repository';
 
 @Component({
   selector: 'app-export-data-modal',
