+++ src/app/elitefour/modals/list-form-modal/list-form-modal.component.ts
@@ -1,6 +1,7 @@
 import {Component, Input, OnInit} from '@angular/core';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
-import {FavoriteList, FavoriteListsRepository, FavoriteListStatus} from '../../backend/favorite-list-interfaces';
+import {FavoriteList, FavoriteListStatus} from '../../backend/favorite-list-interfaces';
+import {FavoriteListsRepository} from '../../backend/favorite-lists-repository';
 
 @Component({
   selector: 'app-add-list-form',
